# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "serve-upvote",
    "description": "Upvote endpoints",
    "license": {
      "name": ""
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/token": {
      "post": {
        "summary": "Generate a token for a grant flow",
        "operationId": "token",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token created"
          },
          "400": {
            "description": "Unauthorized User"
          },
          "404": {
            "description": "Not Found User"
          },
          "500": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/api/v0_noauth/reviews": {
      "get": {
        "summary": "Get Reviews",
        "operationId": "read_many",
        "responses": {
          "200": {
            "description": "Reviews found in database"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TokenRequest": {
        "type": "object",
        "required": [
          "grant_type"
        ],
        "properties": {
          "client_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "optional client ID (as used, for example, in RFC6749's non password non refresh grant flow)",
            "example": null
          },
          "client_secret": {
            "type": [
              "string",
              "null"
            ],
            "description": "optional client secret (as used, e.g., in RFC6749's non (password|refresh) grant flow)",
            "example": null
          },
          "grant_type": {
            "type": "string",
            "description": "RFC6749 grant type",
            "example": "password"
          },
          "password": {
            "type": [
              "string",
              "null"
            ],
            "description": "optional password (as used, for example, in RFC6749's password grant flow)",
            "example": "pass0"
          },
          "username": {
            "type": [
              "string",
              "null"
            ],
            "description": "optional username (as used, for example, in RFC6749's password grant flow)",
            "example": "user0"
          }
        }
      }
    },
    "securitySchemes": {
      "password": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/api/token",
            "scopes": {}
          }
        }
      }
    }
  },
...
}